+++ unix/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(tx)
-
+add_subdirectory(common)
 add_subdirectory(vncconfig)
 add_subdirectory(vncpasswd)
 add_subdirectory(x0vncserver)
