+++ unix/xserver/hw/vnc/XorgGlue.c
@@ -22,7 +22,6 @@
 #endif
 
 #include <assert.h>
-#include <string.h>
 
 #include "scrnintstr.h"
 #ifdef RANDR
@@ -115,253 +114,3 @@ int vncGetScreenY(int scrIdx)
   return screenInfo.screens[scrIdx]->y;
 }
 
-int vncGetScreenWidth(int scrIdx)
-{
-  return screenInfo.screens[scrIdx]->width;
-}
-
-int vncGetScreenHeight(int scrIdx)
-{
-  return screenInfo.screens[scrIdx]->height;
-}
-
-int vncRandRResizeScreen(int scrIdx, int width, int height)
-{
-#ifdef RANDR
-  ScreenPtr pScreen = screenInfo.screens[scrIdx];
-  /* Try to retain DPI when we resize */
-  return RRScreenSizeSet(pScreen, width, height,
-                         pScreen->mmWidth * width / pScreen->width,
-                         pScreen->mmHeight * height / pScreen->height);
-#else
-  return -1;
-#endif
-}
-
-void vncRandRUpdateSetTime(int scrIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  rp->lastSetTime = currentTime;
-#endif
-}
-
-int vncRandRHasOutputClones(int scrIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  for (int i = 0;i < rp->numCrtcs;i++) {
-    if (rp->crtcs[i]->numOutputs > 1)
-      return 1;
-  }
-#endif
-  return 0;
-}
-
-int vncRandRGetOutputCount(int scrIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  return rp->numOutputs;
-#else
-  return 0;
-#endif
-}
-
-int vncRandRGetAvailableOutputs(int scrIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-
-  int availableOutputs;
-  RRCrtcPtr *usedCrtcs;
-  int numUsed;
-
-  int i, j, k;
-
-  usedCrtcs = malloc(sizeof(RRCrtcPtr) * rp->numCrtcs);
-  if (usedCrtcs == NULL)
-    return 0;
-
-  /*
-   * This gets slightly complicated because we might need to hook a CRTC
-   * up to the output, but also check that we don't try to use the same
-   * CRTC for multiple outputs.
-   */
-  availableOutputs = 0;
-  numUsed = 0;
-  for (i = 0;i < rp->numOutputs;i++) {
-    RROutputPtr output;
-
-    output = rp->outputs[i];
-
-    if (output->crtc != NULL)
-      availableOutputs++;
-    else {
-      for (j = 0;j < output->numCrtcs;j++) {
-        if (output->crtcs[j]->numOutputs != 0)
-          continue;
-
-        for (k = 0;k < numUsed;k++) {
-          if (usedCrtcs[k] == output->crtcs[j])
-            break;
-        }
-        if (k != numUsed)
-          continue;
-
-        availableOutputs++;
-
-        usedCrtcs[numUsed] = output->crtcs[j];
-        numUsed++;
-
-        break;
-      }
-    }
-  }
-
-  free(usedCrtcs);
-
-  return availableOutputs;
-#else
-  return 0;
-#endif
-}
-
-char *vncRandRGetOutputName(int scrIdx, int outputIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  return strdup(rp->outputs[outputIdx]->name);
-#else
-  return strdup("");
-#endif
-}
-
-int vncRandRIsOutputEnabled(int scrIdx, int outputIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-
-  if (rp->outputs[outputIdx]->crtc == NULL)
-    return 0;
-  if (rp->outputs[outputIdx]->crtc->mode == NULL)
-    return 0;
-
-  return 1;
-#else
-  return 0;
-#endif
-}
-
-int vncRandRIsOutputUsable(int scrIdx, int outputIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-
-  RROutputPtr output;
-  int i;
-
-  output = rp->outputs[outputIdx];
-  if (output->crtc != NULL)
-    return 1;
-
-  /* Any unused CRTCs? */
-  for (i = 0;i < output->numCrtcs;i++) {
-    if (output->crtcs[i]->numOutputs == 0)
-      return 1;
-  }
-#endif
-
-  return 0;
-}
-
-int vncRandRDisableOutput(int scrIdx, int outputIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  RRCrtcPtr crtc;
-
-  crtc = rp->outputs[outputIdx]->crtc;
-  if (crtc == NULL)
-    return 0;
-
-  return RRCrtcSet(crtc, NULL, crtc->x, crtc->y, crtc->rotation, 0, NULL);
-#else
-  return -1;
-#endif
-}
-
-unsigned int vncRandRGetOutputId(int scrIdx, int outputIdx)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  return rp->outputs[outputIdx]->id;
-#else
-  return 0;
-#endif
-}
-
-void vncRandRGetOutputDimensions(int scrIdx, int outputIdx,
-                            int *x, int *y, int *width, int *height)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-  int swap;
-
-  *x = rp->outputs[outputIdx]->crtc->x;
-  *y = rp->outputs[outputIdx]->crtc->y;
-  *width = rp->outputs[outputIdx]->crtc->mode->mode.width;
-  *height = rp->outputs[outputIdx]->crtc->mode->mode.height;
-
-  switch (rp->outputs[outputIdx]->crtc->rotation & 0xf) {
-  case RR_Rotate_90:
-  case RR_Rotate_270:
-    swap = *width;
-    *width = *height;
-    *height = swap;
-    break;
-  }
-#endif
-}
-
-int vncRandRReconfigureOutput(int scrIdx, int outputIdx, int x, int y,
-                              int width, int height)
-{
-#ifdef RANDR
-  rrScrPrivPtr rp = rrGetScrPriv(screenInfo.screens[scrIdx]);
-
-  RROutputPtr output;
-  RRCrtcPtr crtc;
-  RRModePtr mode;
-
-  int i;
-
-  output = rp->outputs[outputIdx];
-  crtc = output->crtc;
-
-  /* Need a CRTC? */
-  if (crtc == NULL) {
-    for (i = 0;i < output->numCrtcs;i++) {
-      if (output->crtcs[i]->numOutputs != 0)
-        continue;
-
-      crtc = output->crtcs[i];
-      break;
-    }
-
-    /* Couldn't find one... */
-    if (crtc == NULL)
-      return -1;
-  }
-
-  /* Make sure we have the mode we want */
-  mode = vncRandRCreatePreferredMode(output, width, height);
-  if (mode == NULL)
-    return -1;
-
-  /* Reconfigure new mode and position */
-  return RRCrtcSet(crtc, mode, x, y, crtc->rotation, 1, &output);
-#else
-  return -1;
-#endif
-}
