+++ unix/xserver/hw/vnc/XserverDesktop.h
@@ -37,6 +37,7 @@
 #include <rfb/Configuration.h>
 #include <rfb/VNCServerST.h>
 #include <rdr/SubstitutingInStream.h>
+#include <unixcommon.h>
 #include "Input.h"
 
 namespace rfb {
@@ -117,7 +118,6 @@ protected:
   virtual bool handleTimeout(rfb::Timer* t);
 
 private:
-  rfb::ScreenSet computeScreenLayout();
 
   int screenIndex;
   rfb::VNCServerST* server;
@@ -133,7 +133,6 @@ private:
   rfb::Timer queryConnectTimer;
 
 #ifdef RANDR
-  typedef std::map<unsigned int, rdr::U32> OutputIdMap;
   OutputIdMap outputIdMap;
 #endif
